riscv/ld.lld: temporary workaround to get .sdata included in segments mappings
Currently, while linking with ld.lld, any section that comes after .rtemsrwset
(which has a size of 0, hence goes to None) goes to None as well even though
their size isn't 0.

This reordering temporarily fixes this problem to load .sdata in a segment which
contains some symbols that have values.